pyproject.toml
[tool.poetry]
name = "pyAtmosLogger"
version = "0.3"
description = "Python Data Logger for Atmospheric Application"
authors = ["Marcus G. Müller"]
readme = "README.md"
packages = [{include = "pyAtmosLogger"}]
license = "MIT"
homepage = "https://github.com/marcusgmueller/pyAtmosLogger"
repository = "https://github.com/marcusgmueller/pyAtmosLogger"
documentation = "https://pyatmoslogger.readthedocs.io/en/latest/index.html"

[tool.poetry.scripts]
pyAtmosLogger = "pyAtmosLogger.pyAtmosLogger:pyAtmosLogger"

[tool.poetry.dependencies]
python          = "*"
argparse        = "*"
pyyaml          = "*"
xarray          = "*"
pandas          = "*"
pySerial        = "*"
scipy           = "*"
numpy           = "*"
minimalmodbus   = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
